+++ lsassy/exec/__init__.py
@@ -1,0 +1,5 @@
+# Author:
+#  Romain Bentz (pixis - @hackanddo)
+# Website:
+#  https://beta.hackndo.com
+
