+++ lsassy/exec/wmi.py
