+++ lsassy/utils/defines.py
