+++ lsassy/utils/utils.py
@@ -9,7 +9,7 @@ import sys
 import pkg_resources
 from netaddr import IPAddress, IPRange, IPNetwork, AddrFormatError
 
-from lsassy.defines import *
+from lsassy.utils.defines import *
 
 version = pkg_resources.require("lsassy")[0].version
